src/requirements.txt
requests==2.31.0
pandas==2.2.0
more-itertools==8.10.0
python-dateutil==2.8.2